refactor(robot): extract Go2 camera calibration
